Correciones en el modelo de EntitySncf e Indeminizacion

+++ Facturama/Models/Complements/Payroll/Indemnification.cs
@@ -8,7 +8,7 @@ namespace Facturama.Models.Complements.Payroll
 		public decimal TotalPaid { get; set; }
 
 		[JsonProperty("YearsOfService")]
-		public decimal YearsOfService { get; set; }
+		public int YearsOfService { get; set; }
 
 		[JsonProperty("LastMonthlySalaryOrd")]
 		public decimal LastMonthlySalaryOrd { get; set; }
